cleaned ui and controller

+++ dental/routes/web.php
@@ -9,7 +9,10 @@ use App\Http\Controllers\AdminController;
 use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\ServiceController;
 use App\Http\Controllers\AppointmentController;
+use App\Http\Controllers\HomeController;
+use App\Http\Controllers\ReceptionController;
 use App\Models\Appointment as ModelsAppointment;
+use GuzzleHttp\Middleware;
 
 Route::get('/', function () {
     return Inertia::render('Welcome', [
@@ -20,9 +23,7 @@ Route::get('/', function () {
     ]);
 });
 
-Route::get('/dashboard', function () {
-    return Inertia::render('Dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
+Route::get('/dashboard', [HomeController::class,'dashboardSwitch'])->middleware(['auth', 'verified'])->name('dashboard');
 
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
@@ -30,7 +31,7 @@ Route::middleware('auth')->group(function () {
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 });
 
-Route::get('/appointments',[AppointmentController::class,'index'])->name('appointment.index');
+Route::middleware(['auth','role:admin|reception'])->get('/appointments',[AppointmentController::class,'index'])->name('appointment.index');
 
 Route::get('/appointment/create',[AppointmentController::class,'create'])->name('appointment.create');
 
@@ -45,7 +46,7 @@ Route::put('/appointments/{appointment}', [AppointmentController::class, 'update
     ->name('appointments.update');
 
 //delete appointment
-Route::delete('/appointments/{appointment}',[AppointmentController::class,'destroy'])->name('appointment.delete');
+Route::delete('/appointments/{appointment}',[AppointmentController::class,'destroy'])->name('appointment.delete')->middleware('auth','role:admin');
 
 //all the core routes with methods of service created
 Route::resource('services', ServiceController::class);
@@ -53,8 +54,28 @@ Route::resource('services', ServiceController::class);
 //all core crud routes and methods of users
 Route::resource('users',UserController::class);
 
-//reception side
-// Route::get('/reception',[Rece])
+ 
 
-Route::get('/admin',[AdminController::class,'index']);
+Route::get('/admin',[AdminController::class,'index'])->middleware('auth','role:admin');
+
+Route::get('/reception',[ReceptionController::class,'index'])->middleware(['auth','role:reception']);
+
+
+
+Route::get('/debug-role', function() {
+    $user = auth()->user();
+    return [
+        'user' => $user->id ?? null,
+        'roles' => $user ? $user->getRoleNames() : 'not logged in',
+    ];
+});
+
+Route::post('/appointments/{appointment}/checkin', 
+    [AppointmentController::class, 'checkIn'])
+    ->middleware(['role:reception']);
+// Route::get("/reception/overview",[Rec])
+Route::get("/",[HomeController::class,'index']);
+
+Route::get('/dashboardp',[HomeController::class,'dashboard'])->name('patient.dashboard');
+
 require __DIR__.'/auth.php';
